@@ -18,12 +18,15 @@
   },
   "homepage": "https://github.com/soulmatedev/production-project#readme",
   "devDependencies": {
+    "@types/node": "^20.11.20",
+    "@types/webpack": "^5.28.5",
     "html-webpack-plugin": "^5.5.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3",
     "webpack": "^5.69.1",
     "webpack-cli": "^4.9.2"
   },
   "dependencies": {
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.5.5"
+    "ts-loader": "^9.2.6"
   }
 }
